add sending photo feature

+++ application/controllers/Auth.php
@@ -68,36 +68,46 @@ class Auth extends CI_Controller {
         "ticket_solution" => $this->json["solution"],
         "ticket_status" => "SOLVED"
       );
+
       // if update status ticket is success
       if($this->telegram_message_model->change($data, $where)){
         $this->load->library("telegram", array("bot_id" => "331710692:AAGLqH4Yidz7ifiho9EM_y_2xPNfrK3Z-08"));
 
         // get chat id, message id, ticket id, ticket solution, ticket message
-        $telegram_message = $this->telegram_message_model->get($this->json["ticketID"]);
-
-        // send reply to user
-        $content = array(
-					"chat_id" => $telegram_message->CHAT_ID,
-					"text" => "Komplain anda dgn ticket ID #".$telegram_message->TICKET_ID." telah selesai.".chr(10)."Permasalahan: ".$telegram_message->MESSAGE.chr(10)."Solusi: ".$telegram_message->TICKET_SOLUTION,
-					"reply_to_message_id" => $telegram_message->MESSAGE_ID,
-          "parse_mode" => "HTML"
-        );
-        $this->telegram->sendMessage($content);
-
-        // delete data from user_accepted_request
-        $this->load->model("user_accepted_request_model");
         $where = array(
-          "ticket_id" => $telegram_message->TICKET_ID,
-          "username_id" => $telegram_message->USERNAME_ID
+          "ticket_id" => $this->json["ticketID"]
         );
 
-        $this->user_accepted_request_model->destroy($where);
+        $telegram_message = $this->telegram_message_model->get($where);
 
-        //endpushchat
-        $this->load->model("pushchat_model");
-        $this->pushchat_model->destroy($where);
+        if($telegram_message != null){
+          // send reply to user
+          $content = array(
+  					"chat_id" => $telegram_message->CHAT_ID,
+  					"text" => "Komplain anda dgn ticket ID #".$telegram_message->TICKET_ID." telah selesai.".chr(10)."<b>Permasalahan</b>: ".$telegram_message->MESSAGE.chr(10).chr(10)."<b>Solusi:</b> ".$telegram_message->TICKET_SOLUTION,
+  					"reply_to_message_id" => $telegram_message->MESSAGE_ID,
+            "parse_mode" => "HTML"
+          );
+          $this->telegram->sendMessage($content);
 
-        $result = array("result" => array("success" => "Close ticket is successed"));
+          // delete data from user_accepted_request
+          $this->load->model("user_accepted_request_model");
+          $where = array(
+            "ticket_id" => $telegram_message->TICKET_ID,
+            "username_id" => $telegram_message->USERNAME_ID
+          );
+
+          $this->user_accepted_request_model->destroy($where);
+
+          //endpushchat
+          $this->load->model("pushchat_model");
+          $this->pushchat_model->destroy($where);
+
+          $result = array("result" => array("success" => "Close ticket is successed"));
+        }
+        else {
+          $result = array("result" => array("error" => "Close ticket is failed"));
+        }
       }
       else {
         $result = array("result" => array("error" => "Close ticket is failed"));
@@ -117,20 +127,38 @@ class Auth extends CI_Controller {
 
       // get chat id, message id, ticket id
       $this->load->model("telegram_message_model");
-      $telegram_message = $this->telegram_message_model->get($this->json["ticketID"]);
+
+      $where = array(
+        "ticket_id" => $this->json["ticketID"]
+      );
+      $telegram_message = $this->telegram_message_model->get($where);
 
       if($telegram_message != null) {
         $this->load->library("telegram", array("bot_id" => "331710692:AAGLqH4Yidz7ifiho9EM_y_2xPNfrK3Z-08"));
 
-        // send message to user
-        $content = array(
-          "chat_id" => $telegram_message->CHAT_ID,
-          "text" => $this->json["text"],
-          "reply_to_message_id" => $telegram_message->MESSAGE_ID,
-          "parse_mode" => "HTML"
-        );
+        // send message or photo to user
+        if($this->json["type"] == "text") {
+          $content = array(
+            "chat_id" => $telegram_message->CHAT_ID,
+            "text" => $this->json["text"],
+            "reply_to_message_id" => $telegram_message->MESSAGE_ID,
+            "parse_mode" => "HTML"
+          );
+          $this->telegram->sendMessage($content);
+        }
+        else {
+          //get file to local
+          $name_file = $this->json["text"];
+          copy("http://172.18.44.227/portalmac/photo/".$name_file, "internal/".$name_file);
 
-        $this->telegram->sendMessage($content);
+          $content = array(
+            "chat_id" => $telegram_message->CHAT_ID,
+            "photo" => "https://bot.bri.co.id/customercare/internal/".$name_file,
+            "reply_to_message_id" => $telegram_message->MESSAGE_ID,
+          );
+          $this->telegram->sendPhoto($content);
+        }
+
         $result = array("result" => array("success" => "Send message to telegram is success"));
       }
       else {
@@ -159,4 +187,5 @@ class Auth extends CI_Controller {
       $this->json = json_decode($content, true);
     }
   }
+
 }
